--- apps/minting-app/utils/trigonometry.ts
@@ -1,84 +1,0 @@
-import assert from "assert";
-
-export type Point = {
-  x: number;
-  y: number;
-};
-
-export function findClosest(center: Point, radius: number, point: Point) {
-  const vX = point.x - center.x;
-  const vY = point.y - center.y;
-  const magV = Math.sqrt(vX * vX + vY * vY);
-
-  return {
-    x: center.x + (vX / magV) * radius,
-    y: center.y + (vY / magV) * radius,
-  };
-}
-
-export function toAngle(center: Point, { x, y }: Point) {
-  const polar = { x: x - center.x, y: y - center.y };
-  const angle = (Math.atan2(polar.y, polar.x) * 180) / Math.PI;
-
-  return angle;
-
-  let carry;
-  let opposite;
-  let adjacent;
-  if (x >= center.x && y <= center.y) {
-    // first quadradant
-    carry = 0;
-    adjacent = center.y - y;
-    opposite = x - center.x;
-  } else if (x >= center.x && y >= center.y) {
-    // second quadradant
-    carry = Math.PI / 2;
-    adjacent = x - center.x;
-    opposite = y - center.y;
-  } else if (x <= center.x && y >= center.y) {
-    // third quadradant
-    carry = Math.PI;
-    adjacent = y - center.y;
-    opposite = center.x - x;
-  } else {
-    // fourth quadradant
-    carry = Math.PI * 1.5;
-    adjacent = center.x - x;
-    opposite = center.y - y;
-  }
-
-  const radians = carry + Math.atan(opposite / adjacent);
-  const degrees = (radians * 180) / Math.PI;
-  assertDegrees(degrees);
-
-  return degrees;
-}
-
-export function toPosition(center: Point, radius: number, degrees: number) {
-  assertDegrees(degrees);
-  const radians = ((degrees - 90) * Math.PI) / 180.0;
-  return {
-    x: center.x + radius * Math.cos(radians),
-    y: center.y + radius * Math.sin(radians),
-  };
-}
-
-export function dimensions(rect: DOMRect) {
-  const { left, right, top, bottom } = rect;
-
-  const width = right - left;
-  const height = bottom - top;
-  const center = { x: left + width / 2, y: top + height / 2 };
-
-  return { center, width, height };
-}
-
-export function clockwiseDelta(start: number, end: number) {
-  assertDegrees(start);
-  assertDegrees(end);
-  return end - start + (start > end ? 360 : 0);
-}
-
-function assertDegrees(a: number) {
-  assert(a >= 0 && a < 360);
-}
